design: 공통 class 분리, login, register page ul 수정
